LPS-143179 Migrate FDS headless display in `objects-web`
