--- modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/object/entries/frontend/taglib/clay/data/set/data/provider/RelatedModelsDataSetDataProvider.java
@@ -1,135 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.object.web.internal.object.entries.frontend.taglib.clay.data.set.data.provider;
-
-import com.liferay.frontend.taglib.clay.data.Filter;
-import com.liferay.frontend.taglib.clay.data.Pagination;
-import com.liferay.frontend.taglib.clay.data.set.provider.ClayDataSetDataProvider;
-import com.liferay.object.model.ObjectDefinition;
-import com.liferay.object.model.ObjectEntry;
-import com.liferay.object.model.ObjectRelationship;
-import com.liferay.object.related.models.ObjectRelatedModelsProvider;
-import com.liferay.object.related.models.ObjectRelatedModelsProviderRegistry;
-import com.liferay.object.scope.ObjectScopeProvider;
-import com.liferay.object.scope.ObjectScopeProviderRegistry;
-import com.liferay.object.service.ObjectDefinitionLocalService;
-import com.liferay.object.service.ObjectRelationshipLocalService;
-import com.liferay.object.web.internal.object.entries.constants.ObjectEntriesClayDataSetDisplayNames;
-import com.liferay.object.web.internal.object.entries.frontend.taglib.clay.data.set.data.model.RelatedModel;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.search.Sort;
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.vulcan.util.TransformUtil;
-
-import java.util.List;
-
-import javax.servlet.http.HttpServletRequest;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Marco Leo
- */
-@Component(
-	property = "clay.data.provider.key=" + ObjectEntriesClayDataSetDisplayNames.RELATED_MODELS,
-	service = ClayDataSetDataProvider.class
-)
-public class RelatedModelsDataSetDataProvider
-	implements ClayDataSetDataProvider<RelatedModel> {
-
-	@Override
-	public List<RelatedModel> getItems(
-			HttpServletRequest httpServletRequest, Filter filter,
-			Pagination pagination, Sort sort)
-		throws PortalException {
-
-		long objectRelationshipId = ParamUtil.getLong(
-			httpServletRequest, "objectRelationshipId");
-
-		ObjectRelationship objectRelationship =
-			_objectRelationshipLocalService.getObjectRelationship(
-				objectRelationshipId);
-
-		ObjectDefinition objectDefinition =
-			_objectDefinitionLocalService.getObjectDefinition(
-				objectRelationship.getObjectDefinitionId2());
-
-		ObjectRelatedModelsProvider<ObjectEntry> objectRelatedModelsProvider =
-			_objectRelatedModelsProviderRegistry.getObjectRelatedModelsProvider(
-				objectDefinition.getClassName(), objectRelationship.getType());
-
-		ObjectScopeProvider objectScopeProvider =
-			_objectScopeProviderRegistry.getObjectScopeProvider(
-				objectDefinition.getScope());
-
-		long objectEntryId = ParamUtil.getLong(
-			httpServletRequest, "objectEntryId");
-
-		return TransformUtil.transform(
-			objectRelatedModelsProvider.getRelatedModels(
-				objectScopeProvider.getGroupId(httpServletRequest),
-				objectRelationshipId, objectEntryId,
-				pagination.getStartPosition(), pagination.getEndPosition()),
-			objectEntry -> new RelatedModel(
-				objectEntry.getObjectEntryId(), objectEntry.getTitleValue()));
-	}
-
-	@Override
-	public int getItemsCount(
-			HttpServletRequest httpServletRequest, Filter filter)
-		throws PortalException {
-
-		long objectRelationshipId = ParamUtil.getLong(
-			httpServletRequest, "objectRelationshipId");
-
-		ObjectRelationship objectRelationship =
-			_objectRelationshipLocalService.getObjectRelationship(
-				objectRelationshipId);
-
-		ObjectDefinition objectDefinition =
-			_objectDefinitionLocalService.getObjectDefinition(
-				objectRelationship.getObjectDefinitionId2());
-
-		ObjectRelatedModelsProvider<ObjectEntry> objectRelatedModelsProvider =
-			_objectRelatedModelsProviderRegistry.getObjectRelatedModelsProvider(
-				objectDefinition.getClassName(), objectRelationship.getType());
-
-		ObjectScopeProvider objectScopeProvider =
-			_objectScopeProviderRegistry.getObjectScopeProvider(
-				objectDefinition.getScope());
-
-		long objectEntryId = ParamUtil.getLong(
-			httpServletRequest, "objectEntryId");
-
-		return objectRelatedModelsProvider.getRelatedModelsCount(
-			objectScopeProvider.getGroupId(httpServletRequest),
-			objectRelationshipId, objectEntryId);
-	}
-
-	@Reference
-	private ObjectDefinitionLocalService _objectDefinitionLocalService;
-
-	@Reference
-	private ObjectRelatedModelsProviderRegistry
-		_objectRelatedModelsProviderRegistry;
-
-	@Reference
-	private ObjectRelationshipLocalService _objectRelationshipLocalService;
-
-	@Reference
-	private ObjectScopeProviderRegistry _objectScopeProviderRegistry;
-
-}
